test(diagnostics-otel): satisfy ownership proof gates
