+++ extensions/diagnostics-otel/src/service.restart.test.ts
@@ -1,6 +1,32 @@
 import { setTimeout as sleep } from "node:timers/promises";
-import { context, metrics, propagation, trace } from "@opentelemetry/api";
+import {
+  context,
+  diag,
+  DiagLogLevel,
+  metrics,
+  propagation,
+  ROOT_CONTEXT,
+  trace,
+} from "@opentelemetry/api";
 import { logs } from "@opentelemetry/api-logs";
+import { AsyncLocalStorageContextManager } from "@opentelemetry/context-async-hooks";
+import { W3CTraceContextPropagator } from "@opentelemetry/core";
+import {
+  InMemoryLogRecordExporter,
+  LoggerProvider,
+  SimpleLogRecordProcessor,
+} from "@opentelemetry/sdk-logs";
+import {
+  AggregationTemporality,
+  InMemoryMetricExporter,
+  MeterProvider,
+  PeriodicExportingMetricReader,
+} from "@opentelemetry/sdk-metrics";
+import {
+  BasicTracerProvider,
+  InMemorySpanExporter,
+  SimpleSpanProcessor,
+} from "@opentelemetry/sdk-trace-base";
 import {
   resetDiagnosticEventsForTest,
   type DiagnosticTraceContext,
@@ -13,7 +39,7 @@ import {
   startLocalOtlpReceiver,
 } from "../../../test/e2e/qa-lab/runtime/otel-test-support.js";
 import { createDiagnosticsOtelService } from "./service.js";
-import { createOtelContext, emitRealSdkSignals } from "./service.test-helpers.js";
+import { createOtelContext, emitRealSdkSignals, startOtelService } from "./service.test-helpers.js";
 
 const PRELOAD_ENV = "OPENCLAW_OTEL_PRELOADED";
 const OWNERSHIP_ENV_KEYS = [
@@ -23,12 +49,14 @@ const OWNERSHIP_ENV_KEYS = [
   "OTEL_EXPORTER_OTLP_TRACES_PROTOCOL",
   "OTEL_EXPORTER_OTLP_METRICS_PROTOCOL",
   "OTEL_EXPORTER_OTLP_LOGS_PROTOCOL",
+  "OTEL_LOG_LEVEL",
 ] as const;
 const OTEL_GLOBAL_API_KEY = Symbol.for("opentelemetry.js.api.1");
 const OTEL_GLOBAL_LOGS_KEY = Symbol.for("io.opentelemetry.js.api.logs");
 
 type OtelGlobalRegistrations = {
   context?: Parameters<typeof context.setGlobalContextManager>[0];
+  diag?: Parameters<typeof diag.setLogger>[0];
   metrics?: Parameters<typeof metrics.setGlobalMeterProvider>[0];
   propagation?: Parameters<typeof propagation.setGlobalPropagator>[0];
   trace?: Parameters<typeof trace.setGlobalTracerProvider>[0];
@@ -48,7 +76,6 @@ const ORIGINAL_ENV = Object.fromEntries(
   OWNERSHIP_ENV_KEYS.map((key) => [key, process.env[key]]),
 ) as Record<(typeof OWNERSHIP_ENV_KEYS)[number], string | undefined>;
 const ORIGINAL_GLOBALS = { ...registeredOtelGlobals() };
-const ORIGINAL_LOGS = registeredOtelLogs();
 const ORIGINAL_LOGS_PROVIDER = Object.hasOwn(globalThis, OTEL_GLOBAL_LOGS_KEY)
   ? logs.getLoggerProvider()
   : undefined;
@@ -98,6 +125,13 @@ afterEach(() => {
   if (ORIGINAL_LOGS_PROVIDER) {
     logs.setGlobalLoggerProvider(ORIGINAL_LOGS_PROVIDER);
   }
+  diag.disable();
+  if (ORIGINAL_GLOBALS.diag) {
+    diag.setLogger(ORIGINAL_GLOBALS.diag, {
+      logLevel: DiagLogLevel.ALL,
+      suppressOverrideMessage: true,
+    });
+  }
   for (const key of OWNERSHIP_ENV_KEYS) {
     const value = ORIGINAL_ENV[key];
     if (value === undefined) {
@@ -108,6 +142,24 @@ afterEach(() => {
   }
   resetDiagnosticEventsForTest();
 });
+
+function captureOtelDiagnostics(): string[] {
+  const messages: string[] = [];
+  const capture = (...args: unknown[]) => {
+    messages.push(args.map((value) => String(value)).join(" "));
+  };
+  diag.setLogger(
+    {
+      debug: () => {},
+      error: capture,
+      info: () => {},
+      verbose: () => {},
+      warn: capture,
+    },
+    { logLevel: DiagLogLevel.ALL, suppressOverrideMessage: true },
+  );
+  return messages;
+}
 
 test("flushes each private generation and leaves global providers untouched", async () => {
   const receiverA = startLocalOtlpReceiver();
@@ -180,5 +232,146 @@ test("flushes each private generation and leaves global providers untouched", as
     await serviceB.stop?.(ctxB);
     await receiverA.close();
     await receiverB.close();
+  }
+}, 30_000);
+
+test("keeps preloaded host providers live and owned by the host after plugin stop", async () => {
+  releaseOtelGlobals();
+  process.env[PRELOAD_ENV] = "1";
+  const externalContextManager = new AsyncLocalStorageContextManager().enable();
+  const externalPropagator = new W3CTraceContextPropagator();
+  const spanExporter = new InMemorySpanExporter();
+  const tracerProvider = new BasicTracerProvider({
+    spanProcessors: [new SimpleSpanProcessor(spanExporter)],
+  });
+  const metricExporter = new InMemoryMetricExporter(AggregationTemporality.CUMULATIVE);
+  const metricProvider = new MeterProvider({
+    readers: [
+      new PeriodicExportingMetricReader({
+        exporter: metricExporter,
+        exportIntervalMillis: 60_000,
+      }),
+    ],
+  });
+  const logExporter = new InMemoryLogRecordExporter();
+  const loggerProvider = new LoggerProvider({
+    processors: [new SimpleLogRecordProcessor({ exporter: logExporter })],
+  });
+  expect(context.setGlobalContextManager(externalContextManager)).toBe(true);
+  expect(propagation.setGlobalPropagator(externalPropagator)).toBe(true);
+  expect(trace.setGlobalTracerProvider(tracerProvider)).toBe(true);
+  expect(metrics.setGlobalMeterProvider(metricProvider)).toBe(true);
+  logs.setGlobalLoggerProvider(loggerProvider);
+  const hostOwners = {
+    context: registeredOtelGlobals()?.context,
+    logs: logs.getLoggerProvider(),
+    metrics: registeredOtelGlobals()?.metrics,
+    propagation: registeredOtelGlobals()?.propagation,
+    trace: registeredOtelGlobals()?.trace,
+  };
+  const { service, ctx } = await startOtelService({
+    traces: true,
+    metrics: true,
+    logs: false,
+  });
+  const emitHostSignals = (generation: string) => {
+    trace.getTracer("host-preloaded").startSpan(`host-${generation}`).end();
+    metrics
+      .getMeter("host-preloaded")
+      .createCounter("host.preloaded.counter")
+      .add(1, { generation });
+    logs.getLogger("host-preloaded").emit({
+      body: `host-${generation}`,
+      severityText: "INFO",
+    });
+  };
+
+  try {
+    expect({
+      context: registeredOtelGlobals()?.context,
+      logs: logs.getLoggerProvider(),
+      metrics: registeredOtelGlobals()?.metrics,
+      propagation: registeredOtelGlobals()?.propagation,
+      trace: registeredOtelGlobals()?.trace,
+    }).toEqual(hostOwners);
+    emitHostSignals("before-stop");
+    await Promise.all([
+      tracerProvider.forceFlush(),
+      metricProvider.forceFlush(),
+      loggerProvider.forceFlush(),
+    ]);
+
+    const incoming = {
+      traceparent: "00-4bf92f3577b34da6a3ce929d0e0e4736-00f067aa0ba902b7-01",
+    };
+    const extracted = propagation.extract(ROOT_CONTEXT, incoming);
+    const outgoing: Record<string, string> = {};
+    await context.with(extracted, async () => {
+      await new Promise<void>((resolve) => {
+        setImmediate(resolve);
+      });
+      propagation.inject(context.active(), outgoing);
+    });
+    expect(outgoing).toEqual(incoming);
+
+    await service.stop?.(ctx);
+    expect({
+      context: registeredOtelGlobals()?.context,
+      logs: logs.getLoggerProvider(),
+      metrics: registeredOtelGlobals()?.metrics,
+      propagation: registeredOtelGlobals()?.propagation,
+      trace: registeredOtelGlobals()?.trace,
+    }).toEqual(hostOwners);
+    emitHostSignals("after-stop");
+    await Promise.all([
+      tracerProvider.forceFlush(),
+      metricProvider.forceFlush(),
+      loggerProvider.forceFlush(),
+    ]);
+
+    expect(spanExporter.getFinishedSpans().map((span) => span.name)).toEqual([
+      "host-before-stop",
+      "host-after-stop",
+    ]);
+    expect(
+      metricExporter
+        .getMetrics()
+        .flatMap((resourceMetrics) => resourceMetrics.scopeMetrics)
+        .flatMap((scopeMetrics) => scopeMetrics.metrics)
+        .map((metric) => metric.descriptor.name),
+    ).toContain("host.preloaded.counter");
+    expect(logExporter.getFinishedLogRecords().map((record) => record.body)).toEqual([
+      "host-before-stop",
+      "host-after-stop",
+    ]);
+  } finally {
+    await service.stop?.(ctx);
+    await loggerProvider.shutdown();
+    await metricProvider.shutdown();
+    await tracerProvider.shutdown();
+  }
+}, 30_000);
+
+test("leaves OTEL_LOG_LEVEL and the process diagnostic logger under host ownership", async () => {
+  releaseOtelGlobals();
+  const messages = captureOtelDiagnostics();
+  const hostDiagOwner = registeredOtelGlobals()?.diag;
+  process.env.OTEL_LOG_LEVEL = "debug";
+  const receiver = startLocalOtlpReceiver();
+  const port = await receiver.listen();
+  const { service, ctx } = await startOtelService({
+    endpoint: `http://127.0.0.1:${port}`,
+    traces: true,
+  });
+
+  try {
+    await emitRealSdkSignals("diag-owner");
+    await service.stop?.(ctx);
+    expect(registeredOtelGlobals()?.diag).toBe(hostDiagOwner);
+    diag.warn("host diagnostic logger remains active");
+    expect(messages).toContain("host diagnostic logger remains active");
+  } finally {
+    await service.stop?.(ctx);
+    await receiver.close();
   }
 }, 30_000);
